% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{convert2mp3}
\alias{convert2mp3}
\title{Convert to MP3}
\usage{
convert2mp3(path, ...)
}
\arguments{
\item{path}{path to the file to convert}

\item{...}{currently ignored}
}
\description{
Convert to MP3
}
\examples{
\dontrun{
if (require(fs) & require(dplyr)) {
  music <- dir_info("~/Music", recurse = TRUE, type = "file", regexp = ".m4a")
  refl <- music \%>\%
    filter(grepl("Reflections", path))
  convert2mp3(refl$path)
  
}
}
}
