% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_extract.etl_beanumber}
\alias{etl_extract.etl_beanumber}
\title{My ETL functions}
\usage{
\method{etl_extract}{etl_beanumber}(obj, pattern = "hampton", ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{pattern}{regular expression matching file names to be deleted. By default,
this matches filenames ending in \code{.csv} and \code{.zip}.}

\item{...}{arguments passed to methods}
}
\description{
My ETL functions
}
\examples{
\dontrun{
if (require(dplyr)) {
  obj <- etl("beanumber") \%>\%
    etl_create()
}
}
}
