% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtr.R
\name{read_listings}
\alias{read_listings}
\title{Retrieve listings}
\usage{
read_listings(locations = "Northampton, MA", property_type = "land", ...)
}
\arguments{
\item{locations}{character vectors of locations}

\item{property_type}{default "land"}

\item{...}{arguments passed to \code{\link[realtR]{listings}}}
}
\description{
Retrieve listings
}
\examples{
\dontrun{
land <- read_listings()
}
}
