% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_lots.R
\name{read_town_lots}
\alias{read_town_lots}
\alias{download_town_lots}
\title{Read town lots}
\usage{
read_town_lots(pattern = "hampton", cache_dir = getwd(), ...)

download_town_lots(pattern = "hampton", cache_dir = getwd(), ...)
}
\arguments{
\item{pattern}{regular expression matching town names}

\item{cache_dir}{directory where shapefiles are stored}

\item{...}{arguments passed to \code{\link[usethis]{use_zip}}}
}
\description{
Read town lots
}
\examples{
\dontrun{
lots <- read_town_lots("ashfield")
}
\dontrun{
zip_dir <- download_town_lots("ashfield")
}
}
